requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
pytest>=7.4.0
prometheus-client>=0.20.0
flask-limiter>=3.5.0
redis>=5.0.0
SQLAlchemy>=2.0.0
psycopg>=3.1.0
PyYAML>=6.0.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-flask>=0.46b0
